requirements.txt
requests==2.22.0
configparser==4.0.2
feedparser==5.2.1
beautifulsoup4==4.8.2
bs4==0.0.1
azure-storage-blob==12.3.2
azure-identity==1.3.1
applicationinsights==0.11.9
chardet<3.1.0,>=3.0.2
idna<2.9,>=2.5
certifi>=2017.4.17
urllib3!=1.25.0,!=1.25.1,<1.26,>=1.21.1
soupsieve>=1.2
azure-core<2.0.0,>=1.6.0